feat(ui): add @chronogrove/ui workspace package and wire theme
Extract Theme UI theme, ChronogroveThemeProvider, color-mode helpers (SSR + browser), Emotion cache, and portable components into packages/ui with subpath exports. Theme keeps thin shims and consumes the package from gatsby-browser, gatsby-ssr, wrapRootElement, and root-wrapper.

Bump gatsby-theme-chronogrove and @chronogrove/ui to 0.76.0. Document in root/theme README, packages/ui README, CHANGELOG, and .cursorrules.

Made-with: Cursor
