+++ packages/ui/babel.config.cjs
@@ -1,0 +1,6 @@
+module.exports = {
+  presets: [
+    ['@babel/preset-env', { targets: { node: 'current' }, modules: 'commonjs' }],
+    ['@babel/preset-react', { runtime: 'classic' }]
+  ]
+}
