+++ packages/ui/jest.setup.cjs
@@ -1,0 +1,1 @@
+require('@testing-library/jest-dom')
