+++ packages/ui/src/color-mode/normalize.js
@@ -1,0 +1,9 @@
+export function normalizeThemeUiColorMode(mode) {
+  if (mode === 'light') {
+    return 'default'
+  }
+  if (mode === 'dark' || mode === 'default') {
+    return mode
+  }
+  return null
+}
