+++ packages/ui/src/helpers/isDarkMode.js
@@ -1,0 +1,2 @@
+const isDarkMode = colorMode => colorMode === 'dark'
+export default isDarkMode
