+++ packages/ui/src/index.js
@@ -1,0 +1,1 @@
+export { ChronogroveThemeProvider } from './provider.js'
