+++ packages/ui/src/provider.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import { Global } from '@emotion/react'
+import { ThemeUIProvider, InitializeColorMode } from 'theme-ui'
+
+export function ChronogroveThemeProvider({ theme, children }) {
+  return (
+    <ThemeUIProvider theme={theme}>
+      <InitializeColorMode />
+      <Global styles={theme.global} />
+      {children}
+    </ThemeUIProvider>
+  )
+}
