+++ packages/ui/src/skip-nav/SkipNavContent.js
@@ -1,0 +1,16 @@
+import React, { forwardRef } from 'react'
+
+const SkipNavContent = forwardRef(function SkipNavContent(
+  { as: Comp = 'div', id = 'skip-nav-content', children, ...props },
+  forwardedRef
+) {
+  return (
+    <Comp {...props} ref={forwardedRef} id={id} data-skip-nav-content='' tabIndex={-1} style={{ outline: 'none' }}>
+      {children}
+    </Comp>
+  )
+})
+
+SkipNavContent.displayName = 'SkipNavContent'
+
+export default SkipNavContent
