+++ packages/ui/src/skip-nav/index.js
@@ -1,0 +1,2 @@
+export { default as SkipNavLink } from './SkipNavLink.js'
+export { default as SkipNavContent } from './SkipNavContent.js'
