+++ pnpm-workspace.yaml
@@ -1,4 +1,5 @@
 packages:
+  - packages/*
   - theme
   - www.chrisvogt.me
   - www.chronogrove.com
