+++ theme/src/components/root-wrapper.js
@@ -5,21 +5,20 @@ import React, { useEffect, useLayoutEffect } from 'react'
 
 import { logColorModeDebugBanner, logColorModeState } from '../helpers/color-mode-debug'
 import AudioPlayer from './audio-player'
+import {
+  THEME_UI_COLOR_MODE_STORAGE_KEY,
+  RECONCILE_COLOR_MODE_EVENT,
+  resolveChronogroveSurfaceColors
+} from '@chronogrove/ui/color-mode'
 
-// Sync DOM before paint when color mode changes (avoids one frame of dark bg + light root / black text)
 const useIsomorphicLayoutEffect = typeof window !== 'undefined' ? useLayoutEffect : useEffect
-
-const LIGHT_BG = '#fdf8f5'
-const DARK_BG = '#14141F'
-const THEME_UI_COLOR_MODE_KEY = 'theme-ui-color-mode'
-const RECONCILE_COLOR_MODE_EVENT = 'theme-ui-reconcile-color-mode'
 
 const normalizeStored = mode => (mode === 'light' ? 'default' : mode || null)
 
 const getStoredColorMode = () => {
   if (typeof window === 'undefined') return null
   try {
-    return normalizeStored(window.localStorage.getItem(THEME_UI_COLOR_MODE_KEY))
+    return normalizeStored(window.localStorage.getItem(THEME_UI_COLOR_MODE_STORAGE_KEY))
   } catch {
     return null
   }
@@ -31,14 +30,12 @@ const RootWrapper = ({ children }) => {
   const { theme } = useThemeUI()
 
   const normalizedColorMode = colorMode === 'light' ? 'default' : colorMode || 'default'
+  const surface = resolveChronogroveSurfaceColors(theme)
 
-  // Show debug banner once when URL param enables color-mode debug
   useEffect(() => {
     if (typeof window !== 'undefined') logColorModeDebugBanner()
   }, [])
 
-  // On route change, reconcile context from localStorage (e.g. after navigation or hash link).
-  // With a single ThemeUIProvider (from gatsby-plugin-theme-ui), no need to re-sync DOM here.
   useEffect(() => {
     const handler = () => {
       const stored = getStoredColorMode()
@@ -50,12 +47,13 @@ const RootWrapper = ({ children }) => {
     return () => window.removeEventListener(RECONCILE_COLOR_MODE_EVENT, handler)
   }, [normalizedColorMode, setColorMode])
 
-  // Sync root (html class, data attribute, background) from theme context so CSS vars and
-  // fallbacks apply. Uses layout effect so DOM is correct before paint when color mode changes.
   useIsomorphicLayoutEffect(() => {
     if (typeof document === 'undefined') return
     const isDark = normalizedColorMode === 'dark'
-    const bgColorRaw = theme?.rawColors?.background || theme?.colors?.background || (isDark ? DARK_BG : LIGHT_BG)
+    const bgColorRaw =
+      theme?.rawColors?.background ||
+      theme?.colors?.background ||
+      (isDark ? surface.darkBackgroundHex : surface.defaultBackgroundHex)
     const htmlElement = document.documentElement
     Array.from(htmlElement.classList)
       .filter(className => className.startsWith('theme-ui-'))
@@ -64,7 +62,7 @@ const RootWrapper = ({ children }) => {
     htmlElement.setAttribute('data-theme-ui-color-mode', normalizedColorMode)
     htmlElement.style.backgroundColor = bgColorRaw
     logColorModeState(normalizedColorMode, theme, 'RootWrapper')
-  }, [normalizedColorMode, theme?.colors?.background, theme?.rawColors?.background, theme])
+  }, [normalizedColorMode, surface.darkBackgroundHex, surface.defaultBackgroundHex, theme])
 
   return (
     <>
